backend/apps/users/migrations/0001_initial.py
import django.db.models.deletion
import phonenumber_field.modelfields
import uuid
from django.conf import settings
from django.db import migrations, models


class Migration(migrations.Migration):
    """Fase 1: User, Profile sin FK a City, ProfileTag."""

    initial = True
    dependencies = [
        ('auth', '0012_alter_user_first_name_max_length'),
    ]

    operations = [
        migrations.CreateModel(
            name='User',
            fields=[
                ('password', models.CharField(max_length=128, verbose_name='password')),
                ('last_login', models.DateTimeField(blank=True, null=True, verbose_name='last login')),
                ('is_superuser', models.BooleanField(default=False, help_text='Designates that this user has all permissions without explicitly assigning them.', verbose_name='superuser status')),
                ('id', models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True, serialize=False)),
                ('email', models.EmailField(max_length=254, unique=True)),
                ('phone', phonenumber_field.modelfields.PhoneNumberField(blank=True, max_length=128, region=None)),
                ('is_active', models.BooleanField(default=True)),
                ('is_staff', models.BooleanField(default=False)),
                ('is_verified', models.BooleanField(default=False)),
                ('created_at', models.DateTimeField(auto_now_add=True)),
                ('updated_at', models.DateTimeField(auto_now=True)),
                ('groups', models.ManyToManyField(blank=True, help_text='The groups this user belongs to.', related_name='user_set', related_query_name='user', to='auth.group', verbose_name='groups')),
                ('user_permissions', models.ManyToManyField(blank=True, help_text='Specific permissions for this user.', related_name='user_set', related_query_name='user', to='auth.permission', verbose_name='user permissions')),
            ],
            options={'verbose_name': 'usuario', 'verbose_name_plural': 'usuarios', 'ordering': ['-created_at']},
        ),
        migrations.CreateModel(
            name='Profile',
            fields=[
                ('id', models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True, serialize=False)),
                ('name', models.CharField(max_length=120)),
                ('age', models.PositiveSmallIntegerField(blank=True, null=True)),
                ('bio', models.TextField(blank=True)),
                ('photo', models.ImageField(blank=True, upload_to='profiles/')),
                ('industry', models.CharField(blank=True, max_length=80)),
                ('stage', models.CharField(blank=True, choices=[('idea','Idea'),('mvp','MVP'),('early','Early stage'),('growth','Growth'),('scale','Scale'),('executive','Ejecutivo / corporativo'),('investor','Inversor'),('other','Otro')], max_length=20)),
                ('looking_for', models.CharField(blank=True, choices=[('cofounder','Cofundador'),('clients','Clientes'),('investors','Inversores'),('talent','Talento'),('mentors','Mentores'),('network','Expandir red'),('friends','Amigos')], max_length=20)),
                ('linkedin_url', models.URLField(blank=True)),
                ('mbti', models.CharField(blank=True, max_length=4)),
                ('nps_avg', models.FloatField(default=0.0)),
                ('onboarding_complete', models.BooleanField(default=False)),
                ('created_at', models.DateTimeField(auto_now_add=True)),
                ('updated_at', models.DateTimeField(auto_now=True)),
                ('user', models.OneToOneField(on_delete=django.db.models.deletion.CASCADE, related_name='profile', to=settings.AUTH_USER_MODEL)),
            ],
            options={'verbose_name': 'perfil', 'verbose_name_plural': 'perfiles'},
        ),
        migrations.CreateModel(
            name='ProfileTag',
            fields=[
                ('id', models.UUIDField(default=uuid.uuid4, editable=False, primary_key=True, serialize=False)),
                ('tag', models.CharField(max_length=50)),
                ('category', models.CharField(choices=[('interest','Interés'),('skill','Habilidad'),('sector','Sector'),('goal','Objetivo')], max_length=20)),
                ('profile', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, related_name='tags', to='users.profile')),
            ],
            options={'verbose_name': 'tag de perfil', 'verbose_name_plural': 'tags de perfil', 'unique_together': {('profile', 'tag')}},
        ),
    ]
